requirements.txt
Django>=5.2.0
djangorestframework
gunicorn
psycopg2-binary
dj-database-url
